test_x11_requirement: make OS X specific.
